+++ src/main/java/ca/corefacility/bioinformatics/irida/pipeline/results/AnalysisSampleUpdater.java
@@ -2,6 +2,7 @@ package ca.corefacility.bioinformatics.irida.pipeline.results;
 
 import java.util.Collection;
 
+import ca.corefacility.bioinformatics.irida.exceptions.PostProcessingException;
 import ca.corefacility.bioinformatics.irida.model.enums.AnalysisType;
 import ca.corefacility.bioinformatics.irida.model.sample.Sample;
 import ca.corefacility.bioinformatics.irida.model.workflow.analysis.Analysis;
@@ -14,13 +15,12 @@ import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSu
 public interface AnalysisSampleUpdater {
 	/**
 	 * Updates a collection of samples with a particular analysis result.
-	 * 
-	 * @param samples
-	 *            The samples to update.
-	 * @param analysis
-	 *            The {@link AnalysisSubmission} to use for updating.
+	 *
+	 * @param samples  The samples to update.
+	 * @param analysis The {@link AnalysisSubmission} to use for updating.
+	 * @throws PostProcessingException if the updater could not complete its processing
 	 */
-	public void update(Collection<Sample> samples, AnalysisSubmission analysis);
+	public void update(Collection<Sample> samples, AnalysisSubmission analysis) throws PostProcessingException;
 
 	/**
 	 * Gets the {@link AnalysisType} this updater service handles.