+++ src/main/java/ca/corefacility/bioinformatics/irida/pipeline/results/impl/AnalysisSubmissionSampleProcessorImpl.java
@@ -1,12 +1,16 @@
 package ca.corefacility.bioinformatics.irida.pipeline.results.impl;
 
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkNotNull;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
+import ca.corefacility.bioinformatics.irida.exceptions.PostProcessingException;
+import ca.corefacility.bioinformatics.irida.model.enums.AnalysisType;
+import ca.corefacility.bioinformatics.irida.model.sample.Sample;
+import ca.corefacility.bioinformatics.irida.model.workflow.analysis.Analysis;
+import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSubmission;
+import ca.corefacility.bioinformatics.irida.pipeline.results.AnalysisSampleUpdater;
+import ca.corefacility.bioinformatics.irida.pipeline.results.AnalysisSubmissionSampleProcessor;
+import ca.corefacility.bioinformatics.irida.repositories.sample.SampleRepository;
+import ca.corefacility.bioinformatics.irida.service.AnalysisSubmissionService;
+import ca.corefacility.bioinformatics.irida.service.analysis.annotations.RunAsUser;
+import com.google.common.collect.Maps;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,17 +19,12 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.google.common.collect.Maps;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
-import ca.corefacility.bioinformatics.irida.model.enums.AnalysisType;
-import ca.corefacility.bioinformatics.irida.model.sample.Sample;
-import ca.corefacility.bioinformatics.irida.model.workflow.analysis.Analysis;
-import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSubmission;
-import ca.corefacility.bioinformatics.irida.pipeline.results.AnalysisSampleUpdater;
-import ca.corefacility.bioinformatics.irida.pipeline.results.AnalysisSubmissionSampleProcessor;
-import ca.corefacility.bioinformatics.irida.repositories.sample.SampleRepository;
-import ca.corefacility.bioinformatics.irida.service.analysis.annotations.RunAsUser;
-import ca.corefacility.bioinformatics.irida.service.sample.SampleService;
+import static com.google.common.base.Preconditions.checkArgument;
+import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * Updates samples from an {@link AnalysisSubmission} with results from the
@@ -38,22 +37,23 @@ public class AnalysisSubmissionSampleProcessorImpl implements AnalysisSubmission
 
 	private final Map<AnalysisType, AnalysisSampleUpdater> analysisSampleUpdaterMap;
 	private final SampleRepository sampleRepository;
+	private final AnalysisSubmissionService analysisSubmissionService;
 
 	/**
 	 * Builds a new {@link AnalysisSubmissionSampleProcessorImpl}.
-	 * 
-	 * @param sampleRepository
-	 *            The {@link SampleRepository}.
-	 * @param analysisSampleUpdaterServices
-	 *            A list of {@link AnalysisSampleUpdater}s to use for updating
-	 *            samples.
+	 *
+	 * @param sampleRepository              The {@link SampleRepository}.
+	 * @param analysisSubmissionService     The {@link AnalysisSubmissionService}
+	 * @param analysisSampleUpdaterServices A list of {@link AnalysisSampleUpdater}s to use for updating
+	 *                                      samples.
 	 */
 	@Autowired
-	public AnalysisSubmissionSampleProcessorImpl(SampleRepository sampleRepository,
+	public AnalysisSubmissionSampleProcessorImpl(SampleRepository sampleRepository, AnalysisSubmissionService analysisSubmissionService,
 			List<AnalysisSampleUpdater> analysisSampleUpdaterServices) {
 		checkNotNull(analysisSampleUpdaterServices, "assemblySampleUpdaterService is null");
 		this.sampleRepository = sampleRepository;
 		this.analysisSampleUpdaterMap = Maps.newHashMap();
+		this.analysisSubmissionService = analysisSubmissionService;
 
 		for (AnalysisSampleUpdater analysisSampleUpdaterService : analysisSampleUpdaterServices) {
 			AnalysisType analysisType = analysisSampleUpdaterService.getAnalysisType();
@@ -79,7 +79,7 @@ public class AnalysisSubmissionSampleProcessorImpl implements AnalysisSubmission
 	@RunAsUser("#analysisSubmission.getSubmitter()")
 	@Transactional(propagation=Propagation.REQUIRES_NEW)
 	@PreAuthorize("hasPermission(#analysisSubmission, 'canUpdateSamplesFromAnalysisSubmission')")
-	public void updateSamples(AnalysisSubmission analysisSubmission) {
+	public void updateSamples(AnalysisSubmission analysisSubmission) throws PostProcessingException {
 		if (!analysisSubmission.getUpdateSamples()) {
 			logger.trace("Will not update samples from results for submission=" + analysisSubmission);
 		} else {
@@ -95,6 +95,9 @@ public class AnalysisSubmissionSampleProcessorImpl implements AnalysisSubmission
 					.get(analysis.getAnalysisType());
 
 			if (analysisSampleUpdaterService != null) {
+				// re-reading submission to ensure file paths are correct
+				analysisSubmission = analysisSubmissionService.read(analysisSubmission.getId());
+
 				analysisSampleUpdaterService.update(samples, analysisSubmission);
 			} else {
 				logger.debug(