+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/models/datatables/DTProjectMember.java
@@ -2,7 +2,6 @@ package ca.corefacility.bioinformatics.irida.ria.web.models.datatables;
 
 import java.util.Date;
 
-import ca.corefacility.bioinformatics.irida.model.joins.Join;
 import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectUserJoin;
 import ca.corefacility.bioinformatics.irida.model.project.Project;
 import ca.corefacility.bioinformatics.irida.model.user.User;